requirements.txt
Pillow>=10.0.0
pytoshop>=1.1.0
